Extract functionality to add/remove entitlements to/from servers
