+++ java/code/src/com/redhat/rhn/domain/entitlement/test/BaseEntitlementTestCase.java
@@ -17,6 +17,8 @@ package com.redhat.rhn.domain.entitlement.test;
 import com.redhat.rhn.domain.entitlement.Entitlement;
 import com.redhat.rhn.domain.server.Server;
 import com.redhat.rhn.manager.entitlement.EntitlementManager;
+import com.redhat.rhn.manager.system.entitling.SystemEntitlementManager;
+import com.redhat.rhn.manager.system.entitling.SystemEntitler;
 import com.redhat.rhn.testing.BaseTestCaseWithUser;
 import com.redhat.rhn.testing.ServerTestUtils;
 
@@ -43,8 +45,8 @@ public abstract class BaseEntitlementTestCase extends BaseTestCaseWithUser {
         Server traditional = ServerTestUtils.createTestSystem(user);
         Server foreign = ServerTestUtils.createForeignSystem(user, "9999");
 
-        traditional.setBaseEntitlement(EntitlementManager.MANAGEMENT);
-        foreign.setBaseEntitlement(EntitlementManager.FOREIGN);
+        SystemEntitlementManager.INSTANCE.setBaseEntitlement(traditional, EntitlementManager.MANAGEMENT);
+        SystemEntitlementManager.INSTANCE.setBaseEntitlement(foreign, EntitlementManager.FOREIGN);
 
         assertTrue(traditional.getValidAddonEntitlementsForServer().size() > 0);
         assertTrue(foreign.getValidAddonEntitlementsForServer().size() == 0);
@@ -54,8 +56,8 @@ public abstract class BaseEntitlementTestCase extends BaseTestCaseWithUser {
         Server traditional = ServerTestUtils.createTestSystem(user);
         Server foreign = ServerTestUtils.createForeignSystem(user, "9999");
 
-        traditional.setBaseEntitlement(EntitlementManager.MANAGEMENT);
-        foreign.setBaseEntitlement(EntitlementManager.FOREIGN);
+        SystemEntitlementManager.INSTANCE.setBaseEntitlement(traditional, EntitlementManager.MANAGEMENT);
+        SystemEntitlementManager.INSTANCE.setBaseEntitlement(foreign, EntitlementManager.FOREIGN);
 
         assertTrue(ent.isAllowedOnServer(traditional, null));
         assertFalse(ent.isAllowedOnServer(foreign, null));
