+++ java/code/src/com/redhat/rhn/domain/entitlement/test/ContainerBuildHostEntitlementTest.java
@@ -19,6 +19,8 @@ import com.redhat.rhn.domain.entitlement.ContainerBuildHostEntitlement;
 import com.redhat.rhn.domain.server.Server;
 import com.redhat.rhn.domain.server.test.MinionServerFactoryTest;
 import com.redhat.rhn.manager.entitlement.EntitlementManager;
+import com.redhat.rhn.manager.system.entitling.SystemEntitlementManager;
+import com.redhat.rhn.manager.system.entitling.SystemEntitler;
 import com.redhat.rhn.testing.ServerTestUtils;
 
 import com.suse.manager.reactor.utils.ValueMap;
@@ -45,8 +47,8 @@ public class ContainerBuildHostEntitlementTest extends BaseEntitlementTestCase {
         minion.setOs("SLES");
         minion.setRelease("12.2");
 
-        traditional.setBaseEntitlement(EntitlementManager.MANAGEMENT);
-        minion.setBaseEntitlement(EntitlementManager.SALT);
+        SystemEntitlementManager.INSTANCE.setBaseEntitlement(traditional, EntitlementManager.MANAGEMENT);
+        SystemEntitlementManager.INSTANCE.setBaseEntitlement(minion, EntitlementManager.SALT);
 
         assertTrue(ent.isAllowedOnServer(minion));
         assertFalse(ent.isAllowedOnServer(traditional));
@@ -69,7 +71,7 @@ public class ContainerBuildHostEntitlementTest extends BaseEntitlementTestCase {
         grains.put("osmajorrelease", "7");
         assertTrue(ent.isAllowedOnServer(minion, new ValueMap(grains)));
 
-        minion.setBaseEntitlement(EntitlementManager.MANAGEMENT);
+        SystemEntitlementManager.INSTANCE.setBaseEntitlement(minion, EntitlementManager.MANAGEMENT);
         assertFalse(ent.isAllowedOnServer(minion, new ValueMap(grains)));
     }
 }
