+++ java/code/src/com/redhat/rhn/domain/entitlement/test/OSImageBuildHostEntitlementTest.java
@@ -6,6 +6,7 @@ import com.redhat.rhn.domain.entitlement.OSImageBuildHostEntitlement;
 import com.redhat.rhn.domain.server.Server;
 import com.redhat.rhn.domain.server.test.MinionServerFactoryTest;
 import com.redhat.rhn.manager.entitlement.EntitlementManager;
+import com.redhat.rhn.manager.system.entitling.SystemEntitlementManager;
 import com.redhat.rhn.testing.ServerTestUtils;
 
 import com.suse.manager.reactor.utils.ValueMap;
@@ -37,8 +38,8 @@ public class OSImageBuildHostEntitlementTest extends BaseEntitlementTestCase {
         minion.setOs("SLES");
         minion.setRelease("12.2");
 
-        traditional.setBaseEntitlement(EntitlementManager.MANAGEMENT);
-        minion.setBaseEntitlement(EntitlementManager.SALT);
+        SystemEntitlementManager.INSTANCE.setBaseEntitlement(traditional, EntitlementManager.MANAGEMENT);
+        SystemEntitlementManager.INSTANCE.setBaseEntitlement(minion, EntitlementManager.SALT);
 
         assertTrue(ent.isAllowedOnServer(minion));
         assertFalse(ent.isAllowedOnServer(traditional));
@@ -61,7 +62,7 @@ public class OSImageBuildHostEntitlementTest extends BaseEntitlementTestCase {
         grains.put("osmajorrelease", "7");
         assertTrue(ent.isAllowedOnServer(minion, new ValueMap(grains)));
 
-        minion.setBaseEntitlement(EntitlementManager.MANAGEMENT);
+        SystemEntitlementManager.INSTANCE.setBaseEntitlement(minion, EntitlementManager.MANAGEMENT);
         assertFalse(ent.isAllowedOnServer(minion, new ValueMap(grains)));
     }
 }
