+++ java/code/src/com/redhat/rhn/frontend/action/systems/entitlements/test/SystemEntitlementsSubmitActionTest.java
@@ -28,6 +28,9 @@ import com.redhat.rhn.frontend.action.systems.entitlements.SystemEntitlementsSub
 import com.redhat.rhn.frontend.struts.RhnAction;
 import com.redhat.rhn.manager.entitlement.EntitlementManager;
 import com.redhat.rhn.manager.system.SystemManager;
+import com.redhat.rhn.manager.system.entitling.SystemEntitlementManager;
+import com.redhat.rhn.manager.system.entitling.SystemEntitler;
+import com.redhat.rhn.manager.system.entitling.SystemUnentitler;
 import com.redhat.rhn.testing.RhnPostMockStrutsTestCase;
 import com.redhat.rhn.testing.ServerTestUtils;
 
@@ -42,6 +45,9 @@ public class SystemEntitlementsSubmitActionTest extends RhnPostMockStrutsTestCas
                                    "system_entitlements.setToManagementEntitled";
     private static final String UNENTITLED =
                                     "system_entitlements.unentitle";
+
+    private SystemEntitlementManager systemEntitlementManager = SystemEntitlementManager.INSTANCE;
+
 
     /**
      * {@inheritDoc}
@@ -84,7 +90,7 @@ public class SystemEntitlementsSubmitActionTest extends RhnPostMockStrutsTestCas
         ServerFactory.save(server);
         OrgFactory.save(user.getOrg());
         UserFactory.save(user);
-        SystemManager.removeAllServerEntitlements(server);
+        systemEntitlementManager.removeAllServerEntitlements(server);
         assertFalse(SystemManager.hasEntitlement(server.getId(), ent));
 
         /*
@@ -152,7 +158,7 @@ public class SystemEntitlementsSubmitActionTest extends RhnPostMockStrutsTestCas
 
         Server server = ServerTestUtils.createVirtHostWithGuests(user, 1);
 
-        SystemManager.removeServerEntitlement(server, EntitlementManager.VIRTUALIZATION);
+        systemEntitlementManager.removeServerEntitlement(server, EntitlementManager.VIRTUALIZATION);
         ServerGroupTest.createTestServerGroup(user.getOrg(),
                 groupType);
 
@@ -182,7 +188,7 @@ public class SystemEntitlementsSubmitActionTest extends RhnPostMockStrutsTestCas
 
         assertTrue(SystemManager.hasEntitlement(server.getId(),
                                         EntitlementManager.MANAGEMENT));
-        SystemManager.entitleServer(server, ent);
+        systemEntitlementManager.addEntitlementToServer(server, ent);
 
         addRequestParameter("addOnEntitlement", selectKey);
         dispatch(SystemEntitlementsSubmitAction.KEY_REMOVE_ENTITLED, server);
