+++ java/code/src/com/suse/manager/webui/controllers/test/VirtualPoolsControllerTest.java
@@ -19,8 +19,8 @@ import static org.hamcrest.Matchers.containsString;
 import com.redhat.rhn.domain.action.Action;
 import com.redhat.rhn.domain.server.Server;
 import com.redhat.rhn.manager.action.ActionManager;
-import com.redhat.rhn.manager.system.SystemManager;
 import com.redhat.rhn.manager.system.VirtualizationActionCommand;
+import com.redhat.rhn.manager.system.entitling.SystemEntitler;
 import com.redhat.rhn.taskomatic.TaskomaticApi;
 import com.redhat.rhn.testing.ServerTestUtils;
 
@@ -68,7 +68,7 @@ public class VirtualPoolsControllerTest extends BaseControllerTestCase {
                     with(containsString("serverfactorytest")));
         }});
         VirtManager.setSaltService(saltServiceMock);
-        SystemManager.mockSaltService(saltServiceMock);
+        SystemEntitler.INSTANCE.setSaltService(saltServiceMock);
 
         host = ServerTestUtils.createVirtHostWithGuests(user, 1, true);
     }
